Changed type of affected transcript overlap percentage.

+++ Unite.Data/Entities/Genome/Variants/CNV/AffectedTranscript.cs
@@ -7,5 +7,5 @@ public class AffectedTranscript : VariantAffectedFeature<Variant, Transcript>
 {
     public int? Distance { get; set; }
     public int? OverlapBpNumber { get; set; }
-    public int? OverlapPercentage { get; set; }
+    public double? OverlapPercentage { get; set; }
 }
